fix: rename a bunch of imports from config to env
